+++ src/js/core/methods/ApplyFlags.js
@@ -16,7 +16,7 @@ export default class ApplyFlags extends AbstractMethod<'applyFlags'> {
 
         const { payload } = this;
 
-        validateParams(payload, [{ name: 'flags', type: 'number', obligatory: true }]);
+        validateParams(payload, [{ name: 'flags', type: 'number', required: true }]);
 
         this.params = {
             flags: payload.flags,