+++ src/js/core/methods/CipherKeyValue.js
@@ -30,7 +30,7 @@ export default class CipherKeyValue extends AbstractMethod<'cipherKeyValue'> {
         payload.bundle.forEach(batch => {
             // validate incoming parameters for each batch
             validateParams(batch, [
-                { name: 'path', obligatory: true },
+                { name: 'path', required: true },
                 { name: 'key', type: 'string' },
                 { name: 'value', type: 'string' },
                 { name: 'encrypt', type: 'boolean' },