+++ src/js/core/methods/ComposeTransaction.js
@@ -61,8 +61,8 @@ export default class ComposeTransaction extends AbstractMethod<'composeTransacti
         const { payload } = this;
         // validate incoming parameters
         validateParams(payload, [
-            { name: 'outputs', type: 'array', obligatory: true },
-            { name: 'coin', type: 'string', obligatory: true },
+            { name: 'outputs', type: 'array', required: true },
+            { name: 'coin', type: 'string', required: true },
             { name: 'push', type: 'boolean' },
             { name: 'account', type: 'object' },
             { name: 'feeLevels', type: 'array' },