+++ src/js/core/methods/CustomMessage.js
@@ -23,8 +23,8 @@ export default class CustomMessage extends AbstractMethod<'customMessage'> {
 
         // validate incoming parameters
         validateParams(payload, [
-            { name: 'message', type: 'string', obligatory: true },
-            { name: 'params', type: 'object', obligatory: true },
+            { name: 'message', type: 'string', required: true },
+            { name: 'params', type: 'object', required: true },
         ]);
 
         if (payload.messages) {
@@ -74,14 +74,14 @@ export default class CustomMessage extends AbstractMethod<'customMessage'> {
         const { payload } = uiResp;
 
         // validate incoming response
-        validateParams(payload, [{ name: 'message', type: 'string', obligatory: true }]);
+        validateParams(payload, [{ name: 'message', type: 'string', required: true }]);
 
         if (payload.message.toLowerCase() === 'release') {
             // release device
             return response;
         }
         // validate incoming parameters
-        validateParams(payload, [{ name: 'params', type: 'object', obligatory: true }]);
+        validateParams(payload, [{ name: 'params', type: 'object', required: true }]);
 
         // change local params and make another call to device
         this.params.message = payload.message;