+++ src/js/core/methods/EosGetPublicKey.js
@@ -34,7 +34,7 @@ export default class EosGetPublicKey extends AbstractMethod<'eosGetPublicKey'> {
         payload.bundle.forEach(batch => {
             // validate incoming parameters for each batch
             validateParams(batch, [
-                { name: 'path', obligatory: true },
+                { name: 'path', required: true },
                 { name: 'showOnTrezor', type: 'boolean' },
             ]);