+++ src/js/core/methods/GetAccountInfo.js
@@ -59,7 +59,7 @@ export default class GetAccountInfo extends AbstractMethod<'getAccountInfo'> {
         this.params = payload.bundle.map(batch => {
             // validate incoming parameters
             validateParams(batch, [
-                { name: 'coin', type: 'string', obligatory: true },
+                { name: 'coin', type: 'string', required: true },
                 { name: 'descriptor', type: 'string' },
                 { name: 'path', type: 'string' },