+++ src/js/core/methods/NEMGetAddress.js
@@ -48,7 +48,7 @@ export default class NEMGetAddress extends AbstractMethod<'nemGetAddress'> {
         payload.bundle.forEach(batch => {
             // validate incoming parameters for each batch
             validateParams(batch, [
-                { name: 'path', obligatory: true },
+                { name: 'path', required: true },
                 { name: 'address', type: 'string' },
                 { name: 'network', type: 'number' },
                 { name: 'showOnTrezor', type: 'boolean' },