+++ src/js/core/methods/blockchain/BlockchainSetCustomBackend.js
@@ -25,7 +25,7 @@ export default class BlockchainSetCustomBackend extends AbstractMethod<'blockcha
 
         // validate incoming parameters
         validateParams(payload, [
-            { name: 'coin', type: 'string', obligatory: true },
+            { name: 'coin', type: 'string', required: true },
             { name: 'blockchainLink', type: 'object' },
         ]);